fix: code drift across xmp + core docstrings post-v0.2
- xmp.py:474 synthesize_xmp docstring "Path A only" → "Path A and Path B"
  (docstring lagged the implementation that already supports Path B)
- core/__init__.py: list chemigram.core.assertions under the v1.2.0+
  reference-target validation section
- tests/unit/core/test_synthesize.py:60: HistoryEntry.iop_order is
  float | None (not int | None) — test factory hint matched the old
  shape

+++ tests/unit/core/test_synthesize.py
@@ -57,7 +57,7 @@ def _make_history_entry(
     num: int = 0,
     modversion: int = 7,
     params: str = "00000040",
-    iop_order: int | None = None,
+    iop_order: float | None = None,
     multi_name: str = "",
     blendop_params: str = "gz_baseline",
     blendop_version: int = 14,
